# gorget_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gorget API",
    "description": "API to run Gorget scanners.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/analyze/output": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Analyze Output",
        "description": "Analyze an output and return the sanitized output and the results of the scanners",
        "operationId": "submit_analyze_output_analyze_output_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeOutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/scan/output": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Scan Output",
        "description": "Scans an output running scanners in parallel without sanitizing the prompt",
        "operationId": "submit_scan_output_scan_output_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/analyze/prompt": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Analyze Prompt",
        "description": "Analyze a prompt and return the sanitized prompt and the results of the scanners",
        "operationId": "submit_analyze_prompt_analyze_prompt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzePromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/analyze/conversation": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Analyze Conversation",
        "description": "Scan a chat history: the latest user message, recent messages together, tool results and, when enabled, the risk accumulated over the conversation",
        "operationId": "submit_analyze_conversation_analyze_conversation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/scan/conversation": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Scan Conversation",
        "description": "Same checks as /analyze/conversation without the sanitized prompt",
        "operationId": "submit_scan_conversation_scan_conversation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/scan/prompt": {
      "post": {
        "tags": [
          "Analyze"
        ],
        "summary": "Submit Scan Prompt",
        "description": "Scans a prompt running scanners in parallel without sanitizing the prompt",
        "operationId": "submit_scan_prompt_scan_prompt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanPromptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeConversationRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Chat history, oldest first"
          },
          "scanners_suppress": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scanners to suppress",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "AnalyzeConversationRequest"
      },
      "AnalyzeConversationResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Whether the conversation is safe"
          },
          "scanners": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Risk scores of individual checks",
            "description": "Scanner names for the latest user message, Name[window] for recent messages scanned together, Name[tool] for tool results, ConversationRisk for the accumulated risk"
          },
          "sanitized_prompt": {
            "type": "string",
            "title": "Sanitized latest user message"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "scanners",
          "sanitized_prompt"
        ],
        "title": "AnalyzeConversationResponse"
      },
      "AnalyzeOutputRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "output": {
            "type": "string",
            "title": "Model output"
          },
          "scanners_suppress": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scanners to suppress",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "output"
        ],
        "title": "AnalyzeOutputRequest"
      },
      "AnalyzeOutputResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Whether the output is safe"
          },
          "scanners": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Risk scores of individual scanners"
          },
          "sanitized_output": {
            "type": "string",
            "title": "Sanitized output"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "scanners",
          "sanitized_output"
        ],
        "title": "AnalyzeOutputResponse"
      },
      "AnalyzePromptRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "scanners_suppress": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scanners to suppress",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AnalyzePromptRequest"
      },
      "AnalyzePromptResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Whether the prompt is safe"
          },
          "scanners": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Risk scores of individual scanners"
          },
          "sanitized_prompt": {
            "type": "string",
            "title": "Sanitized prompt"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "scanners",
          "sanitized_prompt"
        ],
        "title": "AnalyzePromptResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanOutputRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "output": {
            "type": "string",
            "title": "Model output"
          },
          "scanners_suppress": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scanners to suppress",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "output"
        ],
        "title": "ScanOutputRequest"
      },
      "ScanOutputResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Whether the output is safe"
          },
          "scanners": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Risk scores of individual scanners"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "scanners"
        ],
        "title": "ScanOutputResponse"
      },
      "ScanPromptRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "scanners_suppress": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scanners to suppress",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "ScanPromptRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}